\name{HFMM-package}
\alias{HFMM-package}
\alias{HFMM}
\docType{package}
\title{
\packageTitle{HFMM}
}
\description{
\packageDescription{HFMM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HFMM}
\packageIndices{HFMM}
Overview of the package, including the major functions
}
\author{
\packageAuthor{HFMM}

Maintainer: \packageMaintainer{HFMM}
}
\references{
References
}

\keyword{ keywords here }
\seealso{
Include links and codes here
}
\examples{
Offer example here
}
